@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.0.0-0 <4.0.0"
 
 dependencies:
   flutter:
@@ -28,33 +28,33 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  flutter_launcher_icons: ^0.9.2
-  persistent_bottom_nav_bar: ^4.0.2
-  http: ^0.13.3
+  flutter_launcher_icons: ^0.13.1
+  persistent_bottom_nav_bar: ^5.0.2
+  http: ^1.1.0
   select_form_field: ^2.2.0
-  datetime_picker_formfield: ^2.0.0
-  custom_radio_grouped_button: ^2.0.0
-  cupertino_radio_choice: ^1.0.1+1
-  flutter_form_builder: ^6.0.1
-  dart_json_mapper: ^2.1.2
-  json_serializable: ^4.1.4
+  datetime_picker_formfield_new: ^2.1.0
+  custom_radio_grouped_button: ^2.2.0
+  cupertino_radio_choice: ^1.0.2
+  flutter_form_builder: ^9.1.0
+  dart_json_mapper: ^2.2.9
+  json_serializable: ^6.7.1
   hive_flutter: ^1.1.0
-  uuid: ^3.0.4
-  dropdown_search2: ^0.6.10
-  url_launcher: ^6.0.10
-  percent_indicator: ^3.0.1
-  sn_progress_dialog: ^1.0.3
-  connectivity_plus: ^1.1.0
-  badges: ^2.0.1
-  material_floating_search_bar: ^0.3.6
-  syncfusion_flutter_charts: ^19.3.46
+  uuid: ^4.0.0
+  dropdown_search: ^5.0.6
+  url_launcher: ^6.1.14
+  percent_indicator: ^4.2.3
+  sn_progress_dialog: ^1.1.3
+  connectivity_plus: ^4.0.2
+  badges: ^3.1.2
+  material_floating_search_bar_2: ^0.5.0
+  syncfusion_flutter_charts: ^22.2.11
   bottom_navy_bar: ^6.0.0
-  diacritic: ^0.1.3
+  diacritic: ^0.1.4
 
 
 dev_dependencies:
-  flutter_native_splash: ^1.2.1
-  hive_generator: ^1.1.0
+  flutter_native_splash: ^2.3.2
+  hive_generator: ^2.0.1
   build_runner:
   flutter_test:
     sdk: flutter
