@@ -53,18 +53,16 @@ class GenreAdapter extends TypeAdapter<Genre> {
 // JsonSerializableGenerator
 // **************************************************************************
 
-Genre _$GenreFromJson(Map<String, dynamic> json) {
-  return Genre(
-    uuid: json['uuid'] as String?,
-    name: json['name'] as String,
-    createdAt: json['created_at'] == null
-        ? null
-        : DateTime.parse(json['created_at'] as String),
-    updatedAt: json['updated_at'] == null
-        ? null
-        : DateTime.parse(json['updated_at'] as String),
-  );
-}
+Genre _$GenreFromJson(Map<String, dynamic> json) => Genre(
+      uuid: json['uuid'] as String?,
+      name: json['name'] as String,
+      createdAt: json['created_at'] == null
+          ? null
+          : DateTime.parse(json['created_at'] as String),
+      updatedAt: json['updated_at'] == null
+          ? null
+          : DateTime.parse(json['updated_at'] as String),
+    );
 
 Map<String, dynamic> _$GenreToJson(Genre instance) => <String, dynamic>{
       'uuid': instance.uuid,
