@@ -1,6 +1,5 @@
 import 'package:biospdatabase/Model/Benificiary/Benificiary.dart';
 import 'package:biospdatabase/View/BenificiaryUI/BenificiaryUi.dart';
-import 'package:dart_json_mapper/dart_json_mapper.dart';
 import 'package:flutter/material.dart';
 
 class BenificiaryListTile extends StatelessWidget {
@@ -25,7 +24,7 @@ class BenificiaryListTile extends StatelessWidget {
           ),
           title: benificiary.fullName != null
               ? Text(
-                  "${benificiary.fullName ?? ""}".toLowerCase(),
+                  "${benificiary.fullName ?? ""}",
                   style: TextStyle(
                       fontWeight: FontWeight.w800, color: Colors.grey.shade800),
                 )
